refactor: move seeding logic to AppBuilder and streamline integration test setup
