+++ src/OweMe.Identity.IntegrationTests/Helpers/AppBuilder.cs
@@ -1,50 +1,25 @@
 ﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using OweMe.Identity.Server.Setup;
 using Serilog;
 using Testcontainers.PostgreSql;
-using Xunit.Abstractions;
 
-namespace OweMe.Identity.IntegrationTests;
+namespace OweMe.Identity.IntegrationTests.Helpers;
 
-public class IntegrationTestSetup
-{
-    public App Create()
-    {
-        var builder = WebApplication.CreateBuilder();
-        builder.WebHost.UseUrls("https://[::1]:0");
-        
-        var app = new App
-        {
-            Builder = builder
-        };
-        return app;
-    }
-
-    public static void InitGlobalLogging(ITestOutputHelper testOutputHelper)
-    {
-        Log.Logger = new LoggerConfiguration()
-            .WriteTo.Console()
-            .WriteTo.TestOutput(testOutputHelper)
-            .CreateBootstrapLogger();
-    }
-}
-
-public sealed class App : IAsyncDisposable
+public sealed class AppBuilder : IAsyncDisposable
 {
     public required WebApplicationBuilder Builder { get; set; }
 
     private PostgreSqlContainer? _postgresContainer;
-    
-    public App Configure<T>(Action<T> configure)
+
+    public AppBuilder Configure<T>(Action<T> configure)
         where T : class
     {
         Builder.Services.Configure(configure);
         return this;
     }
 
-    public App WithDatabase()
+    public AppBuilder WithDatabase()
     {
         if (_postgresContainer != null)
         {
@@ -55,12 +30,12 @@ public sealed class App : IAsyncDisposable
         return this;
     }
 
-    public App WithConnectionString(string connectionString)
+    public AppBuilder WithConnectionString(string connectionString)
     {
         Builder.Configuration["ConnectionStrings:DefaultConnection"] = connectionString;
         return this;
     }
-    
+
     public async Task<WebApplication> StartAppAsync()
     {
         if (_postgresContainer is not null)
@@ -68,7 +43,7 @@ public sealed class App : IAsyncDisposable
             await _postgresContainer.StartAsync();
             _ = WithConnectionString(_postgresContainer.GetConnectionString());
         }
-        
+
         var app = Builder.AddIdentityServer().Build();
         app.ConfigurePipeline();
         await app.StartAsync();
